Add inheritance to the IFrameController
